+++ register.js
@@ -1,0 +1,4 @@
+const hooks = require('require-extension-hooks');
+const plugin = require('./src');
+
+hooks('vue').plugin(plugin).push();