+++ src/index.js
@@ -1,127 +1,213 @@
+const {
+  join,
+  dirname,
+} = require('path');
+const {
+  readFileSync,
+} = require('fs');
+const {
+  spawnSync,
+} = require('child_process');
 const compiler = require('vue-template-compiler');
 const transpile = require('vue-template-es2015-compiler');
-const { spawnSync } = require('child_process');
-const { join, dirname } = require('path');
-const { readFileSync } = require('fs');
-const postcss = require('postcss')
-const postcssModules = require('postcss-modules-sync').default
+const postcss = require('postcss');
+const postcssModules = require('postcss-modules-sync').default;
+
 const exportsTarget = '((module.exports.default || module.exports).options || module.exports.default || module.exports)';
+
 const defaultConfig = {
-  transpileTemplates: true
+  transpileTemplates: true,
+  sourceMaps: true,
 };
+
 let globalConfig = defaultConfig;
 
-module.exports = function ({ content, filename, hook }) {
-  function retrieveContent(config) {
-    if (config.attrs.src) {
-      const fullPath = join(dirname(filename), config.attrs.src);
-      return readFileSync(fullPath, 'utf8');
-    } else if (config.content) {
-      return config.content;
+let sourceMapSupport = false;
+
+function retrieveContent(
+  filename,
+  config
+) {
+  if (config.attrs.src) {
+    const fullPath = join(dirname(filename), config.attrs.src);
+    return readFileSync(fullPath, 'utf8');
+  } else if (config.content) {
+    return config.content;
+  }
+
+  return '';
+}
+
+function retrieveAndTranspileContent(
+  filename,
+  config,
+  hook,
+  noTranspileLangs,
+  transpileSpecial
+) {
+  const content = retrieveContent(filename, config);
+  const lang = (config.attrs.lang || '').toLowerCase();
+
+  if (lang && noTranspileLangs.includes(lang) === false) {
+    if (transpileSpecial == null) {
+      return hook(lang, content);
     } else {
-      return '';
+      return transpileSpecial(content, lang);
     }
   }
 
-  function retrieveAndTranspileContent(config, noTranspileLangs, transpileSpecial) {
-    let content = retrieveContent(config)
+  return content;
+}
 
-    let lang = config.attrs.lang
-    if (lang) {
-      lang = lang.toLowerCase()
+function getScriptPart(
+  filename,
+  hook,
+  script
+) {
+  if (!script) {
+    throw new Error(`Unable to read ${filename}: could not find a valid <script> tag`);
+  }
+  return retrieveAndTranspileContent(
+    filename,
+    script,
+    hook,
+    [ 'js', 'javascript' ]
+  );
+}
+
+function transpileTemplateSpecial(
+  content,
+  lang
+) {
+  const result = spawnSync(
+    process.execPath,
+    [
+      join(__dirname, './renderTemplate.js'),
+      lang,
+      content,
+    ],
+    {
+      encoding: 'utf-8',
     }
+  );
 
-    if (lang && !noTranspileLangs.includes(lang)) {
-      content = transpileSpecial != null
-        ? transpileSpecial(content, lang)
-        : hook(lang, content);
-    }
-
-    return content
+  if (result.stderr) {
+    throw new Error(result.stderr);
   }
 
-  function getScriptPart(script) {
-    if (!script) {
-      throw new Error('Unable to read ' + filename + ': could not find a valid <script> tag');
-    }
-    return retrieveAndTranspileContent(script, ['js', 'javascript']);
+  return result.stdout;
+}
+
+function getCompiledTemplate(
+  filename,
+  hook,
+  template
+) {
+  if (!template) {
+    return '';
   }
 
-  function getCompiledTemplate(template) {
+  const content = retrieveAndTranspileContent(
+    filename,
+    template,
+    hook,
+    [ 'html' ],
+    globalConfig.transpileTemplates && transpileTemplateSpecial
+  );
 
-    function transpileTemplateSpecial(content, lang) {
-      const result = spawnSync(
-        process.execPath,
-        [
-          join(__dirname, './renderTemplate.js'),
-          lang,
-          content
-        ],
-        { encoding: 'utf-8' }
+  const compiled = compiler.compile(content, { preserveWhitespace: false });
+  const renderFn = `function(){ ${compiled.render} }`;
+  const staticRenderFns = (compiled.staticRenderFns || [])
+    .map((fn) => `function(){ ${fn} }`)
+    .join(',');
+
+  return [
+    ';',
+    transpile(`${exportsTarget}.render=${renderFn};`),
+    transpile(`${exportsTarget}.staticRenderFns = [ ${staticRenderFns} ];`),
+    `${exportsTarget}.render._withStripped = true;`,
+  ].join('\n');
+}
+
+function getCssModuleComputedProps(
+  filename,
+  hook,
+  styles
+) {
+  if (!styles) {
+    return '';
+  }
+
+  const computedProps = styles
+    .filter((style) => style.module !== undefined && style.module !== false)
+    .map((style) => {
+      const moduleName = (typeof style.module) === 'string' ? style.module : '$style';
+      const content = retrieveAndTranspileContent(
+        filename,
+        style,
+        hook,
+        [ 'css' ]
       );
 
-      if (result.stderr) {
-        throw new Error(result.stderr);
-      }
+      let cssClasses;
+      postcss(postcssModules(
+        {
+          generateScopedName: moduleName + '-[local]',
+          getJSON: json => { cssClasses = json; }
+        })).process(content).css;
+      const cssClassesStr = JSON.stringify(cssClasses);
 
-      return result.stdout;
-    }
+      return `${exportsTarget}.computed.${moduleName} = function(){ return ${cssClassesStr}; };`;
+    });
 
-    if (!template)
-      return ''
-
-    // If there is a template then compile to render functions
-    // This avoids the need for a runtime compilation
-    // ES2015 template compiler to support es2015 features
-    const content = retrieveAndTranspileContent(template, ['html'],
-      globalConfig.transpileTemplates ? transpileTemplateSpecial : null);
-
-    const compiled = compiler.compile(content, { preserveWhitespace: false });
-    const renderFn = `function(){${compiled.render}}`;
-    const staticRenderFns = (compiled.staticRenderFns || [])
-      .map(fn => `function(){${fn}}`)
-      .join(',');
-    return `\n;`
-      + transpile(`${exportsTarget}.render=${renderFn};`)
-      + '\n'
-      + transpile(`${exportsTarget}.staticRenderFns = [${staticRenderFns}];`)
-      + `\n${exportsTarget}.render._withStripped = true;`;
+  if (!computedProps.length) {
+    return '';
   }
 
-  function getCssModuleComputedProps(styles) {
-    if (!styles)
-      return ''
+  return `${exportsTarget}.computed = ${exportsTarget}.computed || {}; ${computedProps.join(' ')}`;
+}
 
-    const computedProps = []
-    for (let i = 0; i < styles.length; i++) {
-      const style = styles[i];
-      if (style.module !== undefined && style.module !== false) {
-        const moduleName = (typeof style.module) === 'string' ? style.module : '$style';
-        const content = retrieveAndTranspileContent(style, ['css']);
-        let cssClasses;
-        postcss(postcssModules(
-          {
-            generateScopedName: moduleName + '-[local]',
-            getJSON: json => { cssClasses = json; }
-          })).process(content).css;
-        const cssClassesStr = JSON.stringify(cssClasses)
-        computedProps.push(`${exportsTarget}.computed.${moduleName} = function(){ return ${cssClassesStr}; };`)
-      }
-    }
-    return (computedProps.length > 0)
-      ? `${exportsTarget}.computed = ${exportsTarget}.computed || {};` + computedProps.join(' ')
-      : '';
+module.exports = ({ content, filename, hook }) => {
+  const {
+    template,
+    script,
+    styles,
+  } = compiler.parseComponent(content, { pad: 'line' });
+
+  if (globalConfig.sourceMaps && sourceMapSupport === false) {
+    require('source-map-support').install({
+      hookRequired: true,
+    });
+    sourceMapSupport = true;
   }
 
-  const { template, script, styles } = compiler.parseComponent(content, { pad: 'line' });
-  const scriptPart = getScriptPart(script);
-  const compliledTemplate = getCompiledTemplate(template)
-  const cssModulesComputedProps = getCssModuleComputedProps(styles)
+  const scriptPart = getScriptPart(
+    filename,
+    hook,
+    script
+  );
 
-  const result = `${scriptPart}\n${compliledTemplate}\n${cssModulesComputedProps}`;
+  const compliledTemplate = getCompiledTemplate(
+    filename,
+    hook,
+    template
+  );
+
+  const cssModulesComputedProps = getCssModuleComputedProps(
+    filename,
+    hook,
+    styles
+  );
+
+  const result = [
+    scriptPart,
+    compliledTemplate,
+    cssModulesComputedProps,
+  ].join('\n');
+
   return { content: result };
 };
 
-module.exports.configure = function (config) {
-  globalConfig = Object.assign({}, defaultConfig, config);
+module.exports.configure = (config) => {
+  globalConfig = Object.assign({}, defaultConfig, globalConfig, config);
 };